# Implement EnqueueExtensions