@codefromthecrypt 2023-07-14
ok I cleaned up the PR so that it is ready to merge if ok (backfilled test coverage and notes).

After this, we can add guest config (e.g. plugin to read yaml or otherwise) and PreScore, so we can implement a realistic, configurable plugin like [nodenumber](https://github.com/kubernetes-sigs/kube-scheduler-simulator/blob/06846c8c8313d3bd4f0fcbd4e533fb3d2f8375a1/simulator/docs/sample/nodenumber/plugin.go#L79C3-L79C56).