@codefromthecrypt 2023-07-14
ok all good. bench should have become a little worse on noop, due to extra hook. Things look correct now.