@codefromthecrypt 2023-07-17
@sanposhiho PTAL I think I got everything (pre-squashed)