@codefromthecrypt 2023-07-17
tomorrow I'll work on pre-score and hopefully config. after that, we should be able to port a basic plugin.